make reducer manage sources internally
